@@ -11,11 +11,12 @@ using ProfilerModels;
 
 namespace Profiler.Controllers
 {
-    [Microsoft.AspNetCore.Authorization.Authorize(Roles = Role.Access.MUST_BE_AUTHENTICATED)]
+    [Authorize(Roles = Role.Access.MUST_BE_AUTHENTICATED)]
     public class HomeController : BaseController
     {
         public HomeController(ControllerHelper controllerHelper) : base(controllerHelper) { }
 
+        [Authorize(Roles = Role.Access.MUST_BE_AUTHENTICATED)]
         public IActionResult Index()
         {
             var viewModelWithUser = _controllerHelper_.ReturnViewModelWithUser(HttpContext);
