@@ -142,7 +142,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _utils_utils_requests__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./utils/utils-requests */ \"./Scripts/utils/utils-requests.js\");\n/* harmony import */ var _components_form_user_login__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/form-user-login */ \"./Scripts/components/form-user-login.js\");\n/* harmony import */ var _components_form_user_add_or_edit__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/form-user-add-or-edit */ \"./Scripts/components/form-user-add-or-edit.js\");\n/* harmony import */ var _components_form_test__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/form-test */ \"./Scripts/components/form-test.js\");\n\r\n\r\n\r\n\r\n\r\n\r\nStartup();\r\n\r\nfunction Startup() {\r\n    console.log(\"Welcome!\\nCompile time:\", new Date().toUTCString());\r\n}\r\n\r\n$(document).on(\"click\", \"#button-logout\", function () {\r\n    Object(_utils_utils_requests__WEBPACK_IMPORTED_MODULE_0__[\"requestLogOut\"])();\r\n})\r\n\r\nnew Vue(\r\n    {\r\n        el: '#app',\r\n        components: {\r\n            'form-test': _components_form_test__WEBPACK_IMPORTED_MODULE_3__[\"formTest\"],\r\n            'form-user-register': formUserRegister,\r\n            'form-user-add-or-edit': _components_form_user_add_or_edit__WEBPACK_IMPORTED_MODULE_2__[\"formUserAddOrEdit\"]\r\n        },\r\n        data: {},\r\n        computed: {},\r\n        methods: {}\r\n    }\r\n);\n\n//# sourceURL=webpack:///./Scripts/site.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _utils_utils_requests__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./utils/utils-requests */ \"./Scripts/utils/utils-requests.js\");\n/* harmony import */ var _components_form_user_login__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/form-user-login */ \"./Scripts/components/form-user-login.js\");\n/* harmony import */ var _components_form_user_add_or_edit__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/form-user-add-or-edit */ \"./Scripts/components/form-user-add-or-edit.js\");\n/* harmony import */ var _components_form_test__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/form-test */ \"./Scripts/components/form-test.js\");\n\r\n\r\n\r\n\r\n\r\n\r\nStartup();\r\n\r\nfunction Startup() {\r\n    console.log(\"Welcome!\\nCompile time:\", new Date().toUTCString());\r\n}\r\n\r\n$(document).on(\"click\", \"#button-logout\", function () {\r\n    Object(_utils_utils_requests__WEBPACK_IMPORTED_MODULE_0__[\"requestLogOut\"])();\r\n})\r\n\r\nnew Vue(\r\n    {\r\n        el: '#app',\r\n        components: {\r\n            'form-test': _components_form_test__WEBPACK_IMPORTED_MODULE_3__[\"formTest\"],\r\n            'form-user-login': _components_form_user_login__WEBPACK_IMPORTED_MODULE_1__[\"formUserLogin\"],\r\n            'form-user-add-or-edit': _components_form_user_add_or_edit__WEBPACK_IMPORTED_MODULE_2__[\"formUserAddOrEdit\"]\r\n        },\r\n        data: {},\r\n        computed: {},\r\n        methods: {}\r\n    }\r\n);\n\n//# sourceURL=webpack:///./Scripts/site.js?");
 
 /***/ }),
 
